requirements.txt
streamlit>=1.32.0
requests>=2.31.0
Pillow>=10.0.0
opencv-python>=4.8.0
chromadb>=0.4.22
pandas>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
google-generativeai>=0.5.0
langgraph>=0.2.0
cryptography>=41.0.0
